move backend_state to common states
