+++ src/kubernetes.py
@@ -11,7 +11,7 @@ from lightkube.models.core_v1 import ServicePort, ServiceSpec
 from lightkube.models.meta_v1 import ObjectMeta
 from lightkube.resources.core_v1 import Node, Service
 
-from state.helpers import InvalidStateError
+from state.common import InvalidStateError
 from state.kubernetes import NodePortState
 
 logger = logging.getLogger(__name__)
