--- src/state/helpers.py
@@ -1,8 +1,0 @@
-# Copyright 2026 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Helpers and common utilities for the ingress-configurator-operator states."""
-
-
-class InvalidStateError(Exception):
-    """Exception raised when the state is invalid."""
